refactor(favorites): Move some code out of favorites and into other packages.
Reorganize tha packages of code added for favorites.
