+++ app/src/main/kotlin/ca/rmen/android/poetassistant/main/common/ui/composables/previews/PopupMenuPreviews.kt
@@ -1,11 +1,11 @@
-package ca.rmen.android.poetassistant.main.favorites.composables.previews
+package ca.rmen.android.poetassistant.main.common.ui.composables.previews
 
 import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import ca.rmen.android.poetassistant.main.favorites.composables.FavoritesPopupMenu
+import ca.rmen.android.poetassistant.main.common.ui.composables.FavoritesPopupMenu
 import ca.rmen.android.poetassistant.settings.Layout
 import ca.rmen.android.poetassistant.theme.AppTheme
