+++ app/src/main/kotlin/ca/rmen/android/poetassistant/main/favorites/ui/composables/previews/FavoriteItemPreviews.kt
@@ -1,10 +1,10 @@
-package ca.rmen.android.poetassistant.main.favorites.composables.previews
+package ca.rmen.android.poetassistant.main.favorites.ui.composables.previews
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import ca.rmen.android.poetassistant.main.favorites.composables.FavoriteItem
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.FavoriteItem
 import ca.rmen.android.poetassistant.settings.Layout
 import ca.rmen.android.poetassistant.theme.AppTheme
