+++ app/src/main/kotlin/ca/rmen/android/poetassistant/wotd/WotdLiveData.kt
@@ -23,7 +23,7 @@ import android.content.Context
 import android.text.format.DateUtils
 import android.util.Log
 import ca.rmen.android.poetassistant.Constants
-import ca.rmen.android.poetassistant.main.favorites.FavoritesRepository
+import ca.rmen.android.poetassistant.main.favorites.data.FavoritesRepository
 import ca.rmen.android.poetassistant.R
 import ca.rmen.android.poetassistant.di.NonAndroidEntryPoint
 import ca.rmen.android.poetassistant.main.dictionaries.ResultListData
