+++ app/src/sharedTest/kotlin/ca/rmen/android/poetassistant/shared/main/FavoritesScreenTest.kt
@@ -16,22 +16,22 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import ca.rmen.android.poetassistant.R
 import ca.rmen.android.poetassistant.main.Tab
-import ca.rmen.android.poetassistant.main.favorites.CreateFavoritesShareUseCase
-import ca.rmen.android.poetassistant.main.favorites.FavoritesRepository
+import ca.rmen.android.poetassistant.main.favorites.usecases.CreateFavoritesShareUseCase
+import ca.rmen.android.poetassistant.main.favorites.data.FavoritesRepository
 import ca.rmen.android.poetassistant.main.favorites.FavoritesScreenViewModel
-import ca.rmen.android.poetassistant.main.favorites.GetProcessTextMenuItemsUseCase
-import ca.rmen.android.poetassistant.main.favorites.OpenExternalAppUseCase
-import ca.rmen.android.poetassistant.main.favorites.ShareUseCase
-import ca.rmen.android.poetassistant.main.favorites.composables.CONFIRM_DELETE_DIALOG_CANCEL_BUTTON_TAG
-import ca.rmen.android.poetassistant.main.favorites.composables.CONFIRM_DELETE_DIALOG_CONFIRM_BUTTON_TAG
-import ca.rmen.android.poetassistant.main.favorites.composables.FAVORITES_HEADER_DELETE_ALL_TAG
-import ca.rmen.android.poetassistant.main.favorites.composables.FAVORITES_SCREEN_CONTENT_EMPTY_TAG
-import ca.rmen.android.poetassistant.main.favorites.composables.FAVORITES_SCREEN_CONTENT_LIST_TAG
-import ca.rmen.android.poetassistant.main.favorites.composables.FAVORITE_ITEM_DICTIONARY_TAG
-import ca.rmen.android.poetassistant.main.favorites.composables.FAVORITE_ITEM_RHYMER_TAG
-import ca.rmen.android.poetassistant.main.favorites.composables.FAVORITE_ITEM_ROW_TAG
-import ca.rmen.android.poetassistant.main.favorites.composables.FAVORITE_ITEM_THESAURUS_TAG
-import ca.rmen.android.poetassistant.main.favorites.composables.FavoritesScreen
+import ca.rmen.android.poetassistant.main.common.usecases.GetProcessTextMenuItemsUseCase
+import ca.rmen.android.poetassistant.main.common.usecases.OpenExternalAppUseCase
+import ca.rmen.android.poetassistant.main.common.usecases.ShareUseCase
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.CONFIRM_DELETE_DIALOG_CANCEL_BUTTON_TAG
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.CONFIRM_DELETE_DIALOG_CONFIRM_BUTTON_TAG
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.FAVORITES_HEADER_DELETE_ALL_TAG
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.FAVORITES_SCREEN_CONTENT_EMPTY_TAG
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.FAVORITES_SCREEN_CONTENT_LIST_TAG
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.FAVORITE_ITEM_DICTIONARY_TAG
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.FAVORITE_ITEM_RHYMER_TAG
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.FAVORITE_ITEM_ROW_TAG
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.FAVORITE_ITEM_THESAURUS_TAG
+import ca.rmen.android.poetassistant.main.favorites.ui.composables.FavoritesScreen
 import ca.rmen.android.poetassistant.settings.Layout
 import ca.rmen.android.poetassistant.settings.SettingsRepository
 import ca.rmen.android.poetassistant.theme.AppTheme
